refactor(aapanel): restructure panel services into modular classes
Split monolithic PanelService into specialized services (DatabaseService, DomainService, etc) for better maintainability
Update related job classes to use new service structure
Remove unused imports and clean up service initialization
